Don’t convert Crowdin exported XLIFFs into native files
There are two kind of XLIFF files that can come from Crowdin: either
the file was uploaded as XLIFF, or it was converted into XLIFF by
Crowdin. The latter can be done with any files, including XLIFF, so
XLIFF files obtained with "Download as XLIFF" in Crowdin web UI can end
up in Poedit and look different from the original XLIFF.

Take care to distinguish these cases when syncing Crowdin XLIFF files
and don't replace an XLIFF-converted-to-XLIFF file with native-XLIFF on
sync unexpectedly, as was previously done.